created perturbation file. result is okay.

+++ 2nd/common_globals.h
@@ -1,13 +1,13 @@
 #ifndef _COMMON_GLOBALS_H_INCLUDED_
 #define _COMMON_GLOBALS_H_INCLUDED_
 
-extern long double alpha;
-extern long double omega;
-extern long double mass;
+extern double alpha;
+extern double omega;
+extern double mass;
 extern int number_of_mesh;
-extern long double epsilon;
-extern long double offset;
-extern long double low_lim;
-extern long double up_lim;
+extern double epsilon;
+extern double offset;
+extern double low_lim;
+extern double up_lim;
 
 #endif
